+++ de.metas.adempiere.adempiere/base/src/main/java/de/metas/product/ProductConstants.java
@@ -1,0 +1,38 @@
+package de.metas.product;
+
+/*
+ * #%L
+ * de.metas.adempiere.adempiere.base
+ * %%
+ * Copyright (C) 2016 metas GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 2 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/gpl-2.0.html>.
+ * #L%
+ */
+
+/**
+ * @author metas-dev <dev@metas-fresh.com>
+ *
+ *         Class of constants to have an organized way of keeping the names of the fields used in process parameters but that are not present in any interface of a table.
+ *
+ *         IMPORTANT: Please, change these if you change the names of such fields.
+ */
+public class ProductConstants
+{
+
+	public final static String TARGET_PRODUCT_ID = "M_Product_Target_ID";
+
+	public final static String TARGET_ORG_ID = "AD_Org_Target_ID";
+}