+++ ReleaseNotes.md
@@ -24,7 +24,8 @@ The actual release notes
 # Upcoming Release
 
 ##Features
-
+ - 09822 Inter-Org Product Mapping Process (104151535488)
+    * allows a user to add and remove product mappings between different organisations
  - 09661 Report "Statistik für monatliche Lagermeldung" - grouping per product and selected attributes (106566269211)
  - 09837 Report Gebinde Übersicht change (103471986337)
  - 09740 ADR revenue report by product categories (101851459609) +it +feature